container: Factor out helper to copy image reference
Prep for some signing work, but this also may be generally
useful in the future.
