+++ lib/src/container/encapsulate.rs
@@ -16,7 +16,7 @@ use std::collections::{BTreeMap, HashMap};
 use std::num::NonZeroU32;
 use std::path::Path;
 use std::rc::Rc;
-use tracing::{instrument, Level};
+use tracing::instrument;
 
 /// Annotation injected into the layer to say that this is an ostree commit.
 /// However, because this gets lost when converted to D2S2 https://docs.docker.com/registry/spec/manifest-v2-2/
@@ -258,9 +258,8 @@ async fn build_impl(
         let tempdir = tempfile::tempdir_in("/var/tmp")?;
         let tempdest = tempdir.path().join("d");
         let tempdest = tempdest.to_str().unwrap();
-        let digestfile = tempdir.path().join("digestfile");
 
-        let src = build_oci(
+        let tempoci = build_oci(
             repo,
             ostree_ref,
             Path::new(tempdest),
@@ -269,19 +268,8 @@ async fn build_impl(
             contentmeta,
         )?;
 
-        let mut cmd = skopeo::new_cmd();
-        tracing::event!(Level::DEBUG, "Copying {} to {}", src, dest);
-        cmd.stdout(std::process::Stdio::null()).arg("copy");
-        cmd.arg("--digestfile");
-        cmd.arg(&digestfile);
-        cmd.args(&[src.to_string(), dest.to_string()]);
-        let proc = super::skopeo::spawn(cmd)?;
-        let output = proc.wait_with_output().await?;
-        if !output.status.success() {
-            let stderr = String::from_utf8_lossy(&output.stderr);
-            return Err(anyhow::anyhow!("skopeo failed: {}\n", stderr));
-        }
-        Some(std::fs::read_to_string(digestfile)?.trim().to_string())
+        let digest = skopeo::copy(&tempoci, dest).await?;
+        Some(digest)
     };
     if let Some(digest) = digest {
         Ok(digest)
